chromium/third_party/blink/web_tests/external/wpt/trusted-types/eval-no-csp-no-tt-default-policy.tentative.https-expected.txt

This is a testharness.js-based test.
FAIL eval of TrustedScript works. assert_equals: expected (number) 2 but got (object) object "1+1"
PASS eval of string works and does not call a default policy.
PASS eval of !TrustedScript and !string works.
Harness: the test ran to completion.
